address crasher with voltometer_driver
was getting this on pepe:

2011-11-04 06:41:49,308 CRITICAL [MainThread] root -- something bad happened
Traceback (most recent call last):
  File "./voltometer_driver.py", line 145, in main
    vd.process_forever()
  File "/home/blalor/xbee_dispatcher/consumer.py", line 271, in process_forever
    self.__xb_frame_chan.stop_consuming()
  File "build/bdist.linux-armv5tel/egg/pika/adapters/blocking_connection.py", line 304, in stop_consuming
    self.basic_cancel(consumer_tag)
  File "build/bdist.linux-armv5tel/egg/pika/channel.py", line 336, in basic_cancel
    self._on_cancel_ok, [spec.Basic.CancelOk])
  File "build/bdist.linux-armv5tel/egg/pika/adapters/blocking_connection.py", line 205, in rpc
    self._on_rpc_complete)
TypeError: 'NoneType' object is not iterable

I think the root cause may have been in consumer.py, using the wrong
connection when creating __rpc_chan. To be pedantic/defensive, I added a lock
around uses of __rpc_chan. Also refactored creation of the
voltometer_driver's sens_dat thread so that the pika connection is done in
the thread, per the documentation.