Add support for requests with multiple replies
ND commands trigger multiple responses, the only command that appears to act
this way. So there's no 1:1 correlation between the command and a reply. This
adds support for "async" XBee TX (RPC) requests to the gateway and consumer.
It's kind of clunky, but it's kind of the best I can do right now. These
changes should not impact other consumers.

The node_discover utility is a very messy demo of how this functionality
works. It will grow into a more full-featured XBee network monitor.

WIP: node discoverability

The XBee command "ND" is one of these.